+++ fastlane/4R255W2M5V.json
@@ -1,0 +1,5 @@
+{
+  "key_id": "4R255W2M5V",
+  "issuer_id": "a2f73e05-c4a8-4223-978e-4b05b1b4312d",
+  "key": "-----BEGIN PRIVATE KEY-----\nMIGTAgEAMBMGByqGSM49AgEGCCqGSM49AwEHBHkwdwIBAQQgYKTfaZ9f5uq7/DH5\nIxN54vntQqZADwdA18qr96f/9BmgCgYIKoZIzj0DAQehRANCAATtlac60GrTf+g1\nldG6zOiXhbwgRs1YnZPBHsYGS9SknukFP6G7MTfJHPJcGI9EIi991FfEN2jJRuR5\ndKBSyNbR\n-----END PRIVATE KEY-----"
+}
