+++ fastlane/Appfile
@@ -1,0 +1,13 @@
+app_identifier([
+  "notitime.noffice.app",
+  "notitime.noffice.design-system.app",
+  "notitime.noffice.home.example",
+  "notitime.noffice.organization.example",
+  "notitime.noffice.mypage.example",
+  "notitime.noffice.signup.example"
+]) # The bundle identifier of your app
+apple_id(["do83430208@gmail.com"]) # Your Apple Developer Portal username
+
+
+# For more information about the Appfile, see:
+#     https://docs.fastlane.tools/advanced/#appfile
