+++ fastlane/Fastfile
@@ -1,0 +1,73 @@
+# This file contains the fastlane.tools configuration
+# You can find the documentation at https://docs.fastlane.tools
+#
+# For a list of all available actions, check out
+#
+#     https://docs.fastlane.tools/actions
+#
+# For a list of all available plugins, check out
+#
+#     https://docs.fastlane.tools/plugins/available-plugins
+#
+
+# Uncomment the line if you want fastlane to automatically update itself
+# update_fastlane
+
+default_platform(:ios)
+
+platform :ios do
+  desc "Fetch development certificates and profiles for notitime.noffice.app and build the app"
+  lane :build_noffice_app do
+    app_identifier = "notitime.noffice.app"
+    match(
+      type: "appstore",
+      app_identifier: app_identifier,
+      keychain_password: ENV["MATCH_KEYCHAIN_PASSWORD"]
+    )
+
+    increment_minor_build_number(
+      xcodeproj: "./Projects/Noffice/Noffice.xcodeproj"
+    )
+      
+    build_app(
+      scheme: "Noffice",
+      workspace: "./Noffice-Workspace.xcworkspace",
+      configuration: "dev",
+      clean: true,
+      export_method: "app-store"
+    )
+    
+    api_key = app_store_connect_api_key(
+        key_id: "4R255W2M5V",
+        issuer_id: "a2f73e05-c4a8-4223-978e-4b05b1b4312d",
+        key_filepath: "./fastlane/AuthKey_4R255W2M5V.p8"
+    )
+
+    upload_to_testflight(
+        api_key: api_key
+    )
+    end
+
+  desc "Fetch development certificates and profiles, and build the app for development"
+  lane :development do
+    sync_code_signing(type: "development")
+    build_app(scheme: "dev")
+  end
+
+  desc "Fetch app store certificates and profiles, and build the app for distribution"
+  lane :appstore do
+    sync_code_signing(type: "appstore")
+    build_app(scheme: "prod")
+  end
+end
+
+private_lane :increment_minor_build_number do |options|
+  project_path = options[:xcodeproj]
+  current_version = get_build_number(xcodeproj: project_path)
+  new_version = (current_version.to_f + 0.1).round(1)
+  puts "Current build number: #{current_version}, new build number: #{new_version}"
+  increment_build_number(
+    xcodeproj: project_path,
+    build_number: new_version.to_s
+  )
+end
