+++ fastlane/Matchfile
@@ -1,0 +1,22 @@
+git_url("https://github.com/Team-Notitime/NOFFICE-iOS-FASTLANE.git")
+
+storage_mode("git")
+
+type("development") # The default type, can be: appstore, adhoc, enterprise or development
+
+force_for_new_devices(true) # Auto recreate provisioning profile when change device count of team
+
+app_identifier([
+  "notitime.noffice.app",
+  "notitime.noffice.design-system.app",
+  "notitime.noffice.home.example",
+  "notitime.noffice.organization.example",
+  "notitime.noffice.mypage.example",
+  "notitime.noffice.signup.example"
+])
+username("do83430208@gmail.com") # Your Apple Developer Portal username
+
+# For all available options run `fastlane match --help`
+# Remove the # in the beginning of the line to enable the other options
+
+# The docs are available on https://docs.fastlane.tools/actions/match
